+++ config/app/config.php
@@ -10,6 +10,6 @@ return [
     'item_subtype_id' => env('ITEM_SUBTYPE_ID'),
     'cookie_user' => env('SESSION_NAME_USER'),
     'cookie_bearer' => env('SESSION_NAME_BEARER'),
-    'version' => '0.4.0',
-    'release_date' => '21st July 2022'
+    'version' => '0.5.0',
+    'release_date' => '22nd July 2022'
 ];